Set `pipefail` to ensure correct exit code
If we don't set `pipefail` the exit code will be that of the `tee`
command which will be successful even if the previous command (the one
doing the work) has failed.

As part of this I needed to use `/bin/bash` instead of `/bin/sh` but
since our base image is Ubuntu - this is not a problem.