Don't overwrite branches with custom buckets.
This commit only overwrites a branch if the input uses the default
bucket for the pipeline. If the bucket is set this means that you're
reaching out into the outputs of other pipelines which will not
necessarily have a branch with the same name on.

Those specifying custom buckets will also need to provide the correct
branch name.